requirements.txt
scikit-learn==1.2.2
joblib==1.2.0
numpy==1.24
lightgbm==4.6
pandas==2.3.3
pyyaml==6.0.3